OpenAI/openai_client.py
# openai_client.py
import os
from dotenv import load_dotenv
from openai import OpenAI

# Load environment variables
load_dotenv()

# Set OpenAI API key
api_key = os.getenv("OPENAI_API_KEY")
if not api_key:
    raise ValueError("No OpenAI API key found. Please set it in the .env file.")

# Initialize the OpenAI client
client = OpenAI(api_key=api_key)
